pyproject.toml
[project]
name = "snake-pipe"
version = "0.1.0"
description = "Python-based ETL pipeline framework"
authors = [
    { name = "Vivek Barsagadey", email = "vivek@example.com" }
]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Database",
    "Topic :: Data Processing",
]

dependencies = [
    "pandas>=2.0.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-ai>=0.7.2",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "openpyxl>=3.1.0",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
warehouse = [
    "snowflake-connector-python>=3.7.0",
    "google-cloud-bigquery>=3.11.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
all = [
    "snake-pipe[warehouse,dev]"
]

[project.urls]
Homepage = "https://github.com/vivekbarsagadey/snake-pipe"
Repository = "https://github.com/vivekbarsagadey/snake-pipe"
Documentation = "https://github.com/vivekbarsagadey/snake-pipe/README.md"
Issues = "https://github.com/vivekbarsagadey/snake-pipe/issues"

[project.scripts]
snake-pipe = "snake_pipe.pipeline:run_pipeline"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["snake_pipe"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--cov=snake_pipe",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "pandas.*",
    "sqlalchemy.*",
    "snowflake.*",
    "google.cloud.*",
    "psycopg2.*",
]
ignore_missing_imports = true
